SC-259 Created a function to show alert in utility. Used that function to check if a user confirms to delete the item or not. Used that in different list views as well as the detail views

+++ src/pages/brand-details/brand-details.ts
@@ -5,6 +5,7 @@ import { Component } from '@angular/core';
 import { NavController, NavParams, ToastController } from 'ionic-angular';
 import { SecurityModule } from '../../infra/security/securityModule';
 import { SecurityAccessRightRepo } from '../../model/securityAccessRightRepo';
+import {Utilities} from "../../utility";
 
 @SecurityModule(SecurityAccessRightRepo.BrandAddEdit)
 @Component({
@@ -19,7 +20,9 @@ export class BrandDetails {
   constructor(public navCtrl: NavController,
     private brandService: BrandService,
     private navParams: NavParams,
-    private toastCtrl: ToastController) {
+    private toastCtrl: ToastController,
+    private utility: Utilities
+  ) {
   }
 
   async ionViewDidLoad() {
@@ -47,6 +50,10 @@ export class BrandDetails {
 
   public async delete() {
     try {
+      const deleteItem = await this.utility.confirmRemoveItem("Do you really want to delete this brand!");
+      if(!deleteItem){
+          return;
+      }
       await this.brandService.delete(this.brand);
       let toast = this.toastCtrl.create({
         message: `Brand '${this.brand.name}' has been deleted successfully!`,
